Wallet

+++ models/UserWallet.js
@@ -18,7 +18,7 @@ const UserWallet = sequelize.define(
       },
     },
     balance: {
-      type: DataTypes.DECIMAL(10, 2),
+      type: DataTypes.BIGINT,
       allowNull: false,
     },
     created_at: {
